% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cglasso.R
\name{cg_cd}
\alias{cg_cd}
\title{Coordinate Gradient Descent with Lasso Penalty}
\usage{
cg_cd(X, b, lambda, tol = 1e-04, maxit = 1000)
}
\arguments{
\item{X}{The design matrix.}

\item{b}{The response vector.}

\item{lambda}{The regularization parameter.}

\item{tol}{Tolerance for convergence.}

\item{maxit}{Maximum number of iterations.}
}
\value{
The estimated coefficient vector.
}
\description{
This function implements the coordinate gradient descent algorithm with lasso penalty.
}
\examples{
# example code
rho = 0.9
n = 200
p = 10
Sigma <- matrix(0, nrow = p, ncol = p)
for (i in 1:p) {
 for (j in 1:p) {
    Sigma[i,j] = rho^abs(i-j)
  }
}
library(MASS)
X <- mvrnorm(n = n, mu = rep(0,p), Sigma = Sigma)
b = c(1, rep(0, p-1))
lambda <- 0.5
cg_cd(X, b, lambda, tol = 1e-4, maxit = 1000)
}
