% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cglasso.R
\name{softThresh}
\alias{softThresh}
\title{Soft Thresholding Function}
\usage{
softThresh(x, lambda)
}
\arguments{
\item{x}{Input vector.}

\item{lambda}{Thresholding parameter.}
}
\value{
Soft thresholded vector.
}
\description{
This function performs soft thresholding.
}
